requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.5.2=py311hecd8cb5_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.01.10=hecd8cb5_0
django=4.1=py311hecd8cb5_0
django-environ=0.4.5=py_1
djangorestframework=3.14.0=pyhd8ed1ab_0
flake8=6.0.0=py311hecd8cb5_0
freetype=2.12.1=hd8bbffd_0
giflib=5.2.1=h6c40b1e_3
jpeg=9e=h6c40b1e_1
krb5=1.19.4=hdba6334_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libedit=3.1.20221030=h6c40b1e_0
libffi=3.4.4=hecd8cb5_0
libpng=1.6.39=h6c40b1e_0
libpq=12.9=h1c9f633_3
libtiff=4.5.0=hcec6c5f_2
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
lz4-c=1.9.4=hcec6c5f_0
mccabe=0.7.0=pyhd3eb1b0_0
ncurses=6.4=hcec6c5f_0
openssl=1.1.1t=hca72f7f_0
pillow=9.4.0=py311hcec6c5f_0
pip=23.0.1=py311hecd8cb5_0
psycopg2=2.9.3=py311hdb2fb19_0
pycodestyle=2.10.0=py311hecd8cb5_0
pyflakes=3.0.1=py311hecd8cb5_0
python=3.11.3=h1fd4e5f_0
pytz=2023.3=pyhd8ed1ab_0
readline=8.2=hca72f7f_0
setuptools=66.0.0=py311hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h6c40b1e_0
sqlparse=0.4.3=py311hecd8cb5_0
tk=8.6.12=h5d9f67b_0
tzdata=2023c=h04d1e81_0
wheel=0.38.4=py311hecd8cb5_0
xz=5.4.2=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
